+++ packages/contracts/contracts/test/BalanceManager.t.sol
@@ -251,7 +251,7 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 1,
                 gasPrice: 0,
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.NONE
+                operationFailureType: OperationFailureType.NONE
             })
         );
 
@@ -284,7 +284,7 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 1,
                 gasPrice: 50,
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.NONE
+                operationFailureType: OperationFailureType.NONE
             })
         );
 
@@ -328,7 +328,7 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 1,
                 gasPrice: 50,
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.NONE
+                operationFailureType: OperationFailureType.NONE
             })
         );
 
@@ -371,7 +371,7 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 1,
                 gasPrice: 50,
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.NONE
+                operationFailureType: OperationFailureType.NONE
             })
         );
 
@@ -438,7 +438,7 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 1,
                 gasPrice: 50,
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.NONE
+                operationFailureType: OperationFailureType.NONE
             })
         );
 
@@ -478,7 +478,7 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 1,
                 gasPrice: 0,
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.NONE
+                operationFailureType: OperationFailureType.NONE
             })
         );
 
@@ -509,7 +509,7 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 1,
                 gasPrice: 0,
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.BAD_ROOT
+                operationFailureType: OperationFailureType.JOINSPLIT_BAD_ROOT
             })
         );
 
@@ -541,7 +541,8 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 1,
                 gasPrice: 0,
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.NF_ALREADY_IN_SET
+                operationFailureType: OperationFailureType
+                    .JOINSPLIT_NF_ALREADY_IN_SET
             })
         );
 
@@ -572,7 +573,7 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 1,
                 gasPrice: 0,
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.JOINSPLIT_NFS_SAME
+                operationFailureType: OperationFailureType.JOINSPLIT_NFS_SAME
             })
         );
 
@@ -603,7 +604,7 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 1,
                 gasPrice: 0, // don't reserve any gas, wallet takes up all
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.NONE
+                operationFailureType: OperationFailureType.NONE
             })
         );
 
@@ -649,7 +650,7 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 1,
                 gasPrice: 0,
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.NONE
+                operationFailureType: OperationFailureType.NONE
             })
         );
 
@@ -682,7 +683,7 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 1,
                 gasPrice: 0,
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.NONE
+                operationFailureType: OperationFailureType.NONE
             })
         );
 
@@ -801,7 +802,7 @@ contract BalanceManagerTest is Test {
                 maxNumRefunds: 2,
                 gasPrice: 0,
                 actions: new Action[](0),
-                joinSplitsFailureType: JoinSplitsFailureType.NONE
+                operationFailureType: OperationFailureType.NONE
             })
         );