+++ packages/sdk/src/primitives/operationDigest.ts
@@ -16,31 +16,6 @@ export function computeOperationDigest(
 function hashOperation(
   op: PreSignOperation | SignedOperation | ProvenOperation
 ): string {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let actionPayload = [] as any;
-  for (const action of op.actions) {
-    actionPayload = ethers.utils.solidityPack(
-      ["bytes", "address", "bytes32"],
-      [
-        actionPayload,
-        action.contractAddress,
-        ethers.utils.keccak256(action.encodedFunction),
-      ]
-    );
-  }
-
-  let refundAssetsPayload = [] as any;
-  for (const encodedAsset of op.encodedRefundAssets) {
-    refundAssetsPayload = ethers.utils.solidityPack(
-      ["bytes", "uint256", "uint256"],
-      [
-        refundAssetsPayload,
-        encodedAsset.encodedAssetAddr,
-        encodedAsset.encodedAssetId,
-      ]
-    );
-  }
-
   let joinSplitsPayload = [] as any;
   for (const joinsplit of op.joinSplits) {
     joinSplitsPayload = ethers.utils.solidityPack(
@@ -80,16 +55,59 @@ function hashOperation(
     [op.refundAddr.h1X, op.refundAddr.h1Y, op.refundAddr.h2X, op.refundAddr.h2Y]
   );
 
+  let refundAssetsPayload = [] as any;
+  for (const encodedAsset of op.encodedRefundAssets) {
+    refundAssetsPayload = ethers.utils.solidityPack(
+      ["bytes", "uint256", "uint256"],
+      [
+        refundAssetsPayload,
+        encodedAsset.encodedAssetAddr,
+        encodedAsset.encodedAssetId,
+      ]
+    );
+  }
+
+  let actionsPayload = [] as any;
+  for (const action of op.actions) {
+    actionsPayload = ethers.utils.solidityPack(
+      ["bytes", "address", "bytes32"],
+      [
+        actionsPayload,
+        action.contractAddress,
+        ethers.utils.keccak256(action.encodedFunction),
+      ]
+    );
+  }
+
+  const gasAssetPayload = ethers.utils.solidityPack(
+    ["uint256", "uint256"],
+    [op.encodedGasAsset.encodedAssetAddr, op.encodedGasAsset.encodedAssetId]
+  );
+
   const payload = ethers.utils.solidityPack(
-    ["bytes", "bytes", "bytes", "bytes", "uint256", "uint256", "uint256"],
     [
-      actionPayload,
+      "bytes",
+      "bytes",
+      "bytes",
+      "bytes",
+      "bytes",
+      "uint256",
+      "uint256",
+      "uint256",
+      "uint256",
+      "uint256",
+    ],
+    [
       joinSplitsPayload,
       refundAddrPayload,
       refundAssetsPayload,
+      actionsPayload,
+      gasAssetPayload,
       op.executionGasLimit,
-      op.gasPrice,
       op.maxNumRefunds,
+      op.gasPrice,
+      op.chainId,
+      op.deadline,
     ]
   );