+++ packages/site/src/pages/index.tsx
@@ -33,9 +33,12 @@ import {
 } from "@nocturne-xyz/frontend-sdk";
 import { WALLET_CONTRACT_ADDRESS } from "../config";
 import { TxModal } from "../components/TxModal";
+import { ethers } from "ethers";
+import { RPC_URL } from "../config/contracts";
 
 const ERC20_ID = 0n;
 const TOKEN_ADDRESS = "0x68B1D87F95878fE05B998F19b66F4baba5De1aed";
+const ONE_DAY_SECONDS = 24 * 60 * 60;
 
 const Container = styled.div`
   display: flex;
@@ -156,6 +159,13 @@ const Index = () => {
         [TOKEN_ADDRESS, amount]
       );
 
+    // Kludge, in real frontend you'd provide chainid and deadline not using wallet
+    const provider = new ethers.providers.JsonRpcProvider(RPC_URL);
+    const chainId = BigInt((await provider.getNetwork()).chainId);
+    const deadline = BigInt(
+      (await provider.getBlock("latest")).timestamp + ONE_DAY_SECONDS
+    );
+
     const builder = new OperationRequestBuilder();
     const operationRequest = builder
       .unwrap(asset, amount)
@@ -163,6 +173,8 @@ const Index = () => {
       .refundAsset(asset)
       .maxNumRefunds(5n)
       .gas({ executionGasLimit: 500_000n, gasPrice: 0n })
+      .chainId(chainId)
+      .deadline(deadline)
       .build();
 
     console.log("Operation request: ", operationRequest);